Consolidate validation utilities
