+++ security/__init__.py
@@ -1,8 +1,13 @@
 """Security package exposing validation utilities."""
 
-from .auth_service import SecurityService
-from core.input_validation import InputValidator, Validator
-from .file_validator import SecureFileValidator
+from core.security import InputValidator
+from typing import Protocol
+
+
+class Validator(Protocol):
+    def validate(self, data: str) -> str:
+        ...
+
 from .dataframe_validator import DataFrameSecurityValidator
 from .sql_validator import SQLInjectionPrevention, SQLSecurityConfig
 from .xss_validator import XSSPrevention
@@ -14,10 +19,8 @@ from .validation_exceptions import SecurityViolation
 from .secrets_validator import SecretsValidator, register_health_endpoint
 
 __all__ = [
-    "SecurityService",
     "InputValidator",
     "Validator",
-    "SecureFileValidator",
     "DataFrameSecurityValidator",
     "SQLSecurityConfig",
     "SQLInjectionPrevention",
